requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
openpyxl==3.1.2
aiohttp==3.9.1
